+++ src/App.tsx
@@ -1,9 +1,10 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 
+import Home from './pages/home/home';
+
 import scss from './App.module.scss';
 
 import '~styles/base.scss';
-import Home from './pages/home/';
 
 function App() {
   return (
