+++ src/pages/home/home.module.scss
@@ -1,0 +1,32 @@
+/* normalize.css */
+/* Zera margens e preenchimentos */
+body, h1, h2, h3, h4, h5, h6, p, blockquote, pre, dl, dd, ol, ul, figure, hr {
+  margin: 0;
+  padding: 0;
+}
+
+/* Define a fonte padrão para evitar diferenças entre navegadores */
+body {
+  font-family: Arial, sans-serif;
+}
+
+/* Define o tamanho da caixa de elementos para incluir preenchimento e borda */
+*, ::before, ::after {
+  box-sizing: border-box;
+}
+
+.Home {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgb(117, 31, 245);
+  color: #f2f2f2;
+}
+
+.Logo {
+  border-radius: 50%;
+  width: 30px;
+  height: 30px;
+  margin-left: 10px;
+}
+
